initial rename + timeout bump
